Add distributed test infrastructure
